Working Position Control. Problem: Encoder cannot read when motor turns backwards

+++ WholeProgram/PWM_Encoder_Sensors.h
@@ -57,7 +57,7 @@ typedef struct {
     int positionTicks;
     float Position_Motor; 
     float targetPosition;
-    int RPM; 
+    float RPM; 
 } Encoder_t;
 extern volatile Encoder_t Encoder;
